+++ src/Filters/Has.php
@@ -19,34 +19,63 @@ declare(strict_types=1);
 
 namespace LaravelJsonApi\Eloquent\Filters;
 
+use LaravelJsonApi\Contracts\Schema\Schema;
 use LaravelJsonApi\Eloquent\Contracts\Filter;
+use LaravelJsonApi\Eloquent\Filters\Concerns\HasRelation;
 use LaravelJsonApi\Eloquent\Filters\Concerns\IsSingular;
+use function filter_var;
 
 class Has implements Filter
 {
+    use HasRelation;
     use IsSingular;
-    use Having;
 
     /**
-     * Apply the filter to the query.
+     * Create a new filter.
      *
-     * @param \Illuminate\Database\Eloquent\Builder $query
-     * @param mixed $value
-     * @return \Illuminate\Database\Eloquent\Builder
+     * @param Schema $schema
+     * @param string $fieldName
+     * @param string|null $key
+     * @return static
      */
-    public function apply($query, $value)
+    public static function make(Schema $schema, string $fieldName, string $key = null)
     {
-        $deserializedBoolean = (bool) $this->deserialize($value);
-
-        if ($deserializedBoolean === true) {
-            return $query->has($this->relationName());
-        }
-
-        return $query->whereDoesntHave($this->relationName());
+        return new static($schema, $fieldName, $key);
     }
 
     /**
-     * @param $value
+     * Has constructor.
+     *
+     * @param Schema $schema
+     * @param string $fieldName
+     * @param string|null $key
+     */
+    public function __construct(Schema $schema, string $fieldName, string $key = null)
+    {
+        $this->schema = $schema;
+        $this->fieldName = $fieldName;
+        $this->key = $key;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function apply($query, $value)
+    {
+        $value = $this->deserialize($value);
+        $relationName = $this->relationName();
+
+        if (true === $value) {
+            return $query->has($relationName);
+        }
+
+        return $query->doesntHave($relationName);
+    }
+
+    /**
+     * Deserialize the value.
+     *
+     * @param mixed $value
      * @return bool
      */
     protected function deserialize($value): bool